net: mvpp2: allow to assign a reserved memory region for DMA allocations
This patch adds the possibility to reserve a DMA region in the device
tree, and to assign this memory region to the PPv2 device for DMA
allocations. This helps to avoid cases where the default CMA pool is to
small in the kernel, by having a PPv2 specific pool we can control.

Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>