crypto: inside-secure - ensure buffers match DMA engine requirements
This patch rework how the two buffers used by the engine through DMA
access and controlled by the driver are created. The Inside Secure
engine requires that the DMA buffers it access to retrieve or store data
are aligned on 64 bits. While this worked by chance until now, we saw
cases where the requirement wasn't fulfilled anymore, leading to invalid
hash results.

This patch introduce two DMA pools, configured with the correct
alignment, to store the hash digests and the cached blocks in-between
updates. The driver is reworked to make use of these two pools.

As calling dma_pool_alloc() on each update() is not a good idea in terms
of performances (and stability as memory allocation errors could
happen), the buffers used by both the driver and the engine are moved in
the request context space and allocated in cra_init.

Signed-off-by: Antoine Tenart <antoine.tenart@free-electrons.com>