irqchip/irq-mvebu-icu: support ICU subnodes
Introduce new bindings for the ICU.

Each DT subnode of the ICU represents a type of interrupt that should
be handled separately. Add the possibility for the ICU to have subnodes
and probe each of them automatically with devm_platform_populate(). If
the node as no child, the probe function for NSRs will still be called
'manually'.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>