net: mvpp2: Move PPv2 network driver to a seperate directory
The Marvell PPv2 driver keeps growing, and its single-file structure is
starting to get difficult to navigate into and maintain. This commit
moves the file to dedicated directory, leading the way to organize this
driver into multiple files.

To prepare for a new driver made of multiple files, the mvpp2.c file is
renamed 'mvpp2_main.c", mvpp2 being the name of the driver itself.

This file contains all previous mvpp2 code, except for the macros and
definitions, that are moved to a header file.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>