net: mvpp2: Introduce new file dedicated to Header Parser
Marvell PPv2 controller has a complex Packet Header Parser, that is used
to implement a lot of features. This is used to implement VLAN matching,
VID filtering, unicast/multicast filtering, some levels of flow control
and can be used for a lot more features such as classification.

For now, this represents quite a significant portion of the mvpp2 driver
(about 2500 lines), and having a separate file dealing with this Header
Parser would make sense at this point.

Among other things, this Header Parser has a TCAM filter that can match
various patterns in headers. The structure of the TCAM is defines by a
bunch of includes, that sets all ranges reserved for features like VLAN
/ MAC filtering, and defines all static entries. This is now moved in a
header file associated with parser functions.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>