net: mvpp2: Add hit counter stat for Header Parser entries
One helpfulm feature to help debug the Header Parser TCAM filter in PPv2
is to be able to see if the entries did match something when a packet
comes in. This can be done by using the built-in hit counter for TCAM
entries.

This commit implements reading the counter, and exposing its value on
debugfs for each filter entry. The value resets after reading.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>