DO NOT UPSTREAM: net: mvpp2: stop the wake threshold to rely on the Tx queue size
This patch reworks the Tx queue wake threshold so that it does not rely
on the Tx queue size. Doing so led the gap between the stop and the wake
threshold to increase with the Tx queue size, which meant more resources
where needed, as well as extra latency. This should have nothing to do
with the Tx queue size.

Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>