crypto: inside-secure - split the algorithms list by engine version
This patch splits the common algorithms list by engine version, and puts
it in the device data. Minor reworks are also done as the device data
was already used to store information about the version of the engine
used. This is needed for sha384/512 support as those algorithms will be
supported by the eip197, but not the eip97.

Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>