net: mvpp2: Add RSS support to PPv2 driver
Marvell PPv2 Controller has a hardware classifier, that can be
configured to perform Receive Side Scaling.

Although the driver already has the infrastructure to access RSS tables,
they are actually not used unless we setup specific flows inside the
Classifier.

RSS only works when some flags are attached to the packet, such as the
'SoftwareForwarding' flag that indicates that this packet will be
processed by the CPU, opposite to some hardware re-routing of the
packet.

To do that, a simple flow is created on each interface, that performs
a lookup in the 'C2' engine (one of the many classification engines in
PPv2) to set the correct tags on the packet.

This commits adds a basic default hash generation method, that only uses
the Source and Destination IPv4 address.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>