net: mvpp2: Allow changing RSS Hash parameters
Hash generation engines for RSS in mvpp2 can be configure to use any
combination of fields to compute the hash.

The hash is calculated based on a Header Extracted Key, which is a
concatenation of several fields extracted from the packet header.

There are 2 hash engines in this controller : C3HA and C3HB.

 - C3HA takes up to 4 fields from the header provided they don't contain
   L4 related informations. It's mainly used to perform 2-tuples hashes.

 - C3HB takes up to 4 fields from the header, and automatically adds L4
   info provided by the Header Parser. It can therefore perform up to
   5-tuples hashes.

This commit adds the required ethtool ops to configure these hashes on a
per-port basis.

For now, we only configure a global hash generation procedure per
interface, that will apply to any flow. The flow-type parameter from
ethtool is simply ignored.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>